# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: EasyEscrow.ai API
  version: 1.0.0
  description: Non-custodial atomic swap platform for NFTs and SOL on Solana, plus institutional cross-border USDC escrow payments. Browse the Documentation section for guides, or jump to API Reference for endpoint details.
  contact:
    name: EasyEscrow.ai Support
    email: support@easyescrow.ai
  license:
    name: ISC
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://staging-api.easyescrow.ai
    description: Staging environment (Devnet)
  - url: https://api.easyescrow.ai
    description: Production (Mainnet)
...
paths:
...
  /api/swaps/offers:
    get:
      tags:
        - Atomic Swaps
      summary: List offers
      description: Retrieve swap offers with optional filters and pagination
      parameters:
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/OfferStatus'
          description: Filter by offer status
        - in: query
          name: makerWallet
          schema:
            type: string
          description: Filter by maker wallet address
        - in: query
          name: takerWallet
          schema:
            type: string
          description: Filter by taker wallet address
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Items per page
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Pagination offset
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      offers:
                        type: array
                        items:
                          $ref: '#/components/schemas/OfferSummary'
                      total:
                        type: integer
                      limit:
                        type: integer
                      offset:
                        type: integer
                  timestamp:
                    type: string
                    format: date-time
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    post:
      tags:
        - Atomic Swaps
      summary: Create offer (unified endpoint)
      description: |
        Create a new swap offer with automatic type detection.

        This unified endpoint supports **three offer types**:

        ### 1. Atomic Swap (default)
        Standard swaps for NFT↔NFT, NFT↔SOL, cNFT↔SOL.
        Use `makerWallet`, `offeredAssets`, `requestedAssets`, etc.

        ### 2. cNFT Bid (auto-detected when bidderWallet + targetAssetId + offerLamports present)
        Make a SOL bid on a specific cNFT.
        Use `bidderWallet`, `targetAssetId`, `offerLamports`.

        ### 3. Bundle Two-Phase Swap (auto-detected for 3+ cNFTs or 5+ total assets)
        Complex multi-asset swaps using lock/settle pattern.
        Use `partyA`, `assetsA`, `assetsB`, or standard fields.

        **Field Aliases Supported:**
        - `partyA` → `makerWallet`
        - `partyB` → `takerWallet`
        - `assetsA` → `offeredAssets`
        - `assetsB` → `requestedAssets`
        - `solAmountA` → `offeredSol`
        - `solAmountB` → `requestedSol`

        **Rate Limited**: 10 requests per 15 minutes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOfferRequest'
            examples:
              nftForSol:
                summary: NFT ↔ SOL (Atomic)
                value:
                  makerWallet: "498GViCLvzbGnRoByJCAj7skXkAe3NBpCY2Wghcd2e4R"
                  takerWallet: "8LL197pziojWHtS3zeyJonrh1swKvMZpumfesVmDgUcZ"
                  offeredAssets:
                    - mint: "7xKXtg2CW87d97TXJSDpbD5jBkheTqA83TZRuJosgAsU"
                      isCompressed: false
                  requestedAssets: []
                  offeredSol: "0"
                  requestedSol: "1000000000"
              cnftForSol:
                summary: cNFT ↔ SOL (Atomic)
                value:
                  makerWallet: "498GViCLvzbGnRoByJCAj7skXkAe3NBpCY2Wghcd2e4R"
                  offeredAssets:
                    - mint: "DRiP2Pn2K6fuMLKQmt5rZWqHheXMyUtCeXhe8kDQdxRu"
                      isCompressed: true
                  requestedAssets: []
                  offeredSol: "0"
                  requestedSol: "2000000000"
              cnftBid:
                summary: cNFT Bid (auto-detected)
                value:
                  bidderWallet: "498GViCLvzbGnRoByJCAj7skXkAe3NBpCY2Wghcd2e4R"
                  targetAssetId: "DRiP2Pn2K6fuMLKQmt5rZWqHheXMyUtCeXhe8kDQdxRu"
                  offerLamports: "1000000000"
                  durationSeconds: 86400
              bulkSwapPartyStyle:
                summary: Bundle Swap (partyA/B style)
                value:
                  partyA: "498GViCLvzbGnRoByJCAj7skXkAe3NBpCY2Wghcd2e4R"
                  partyB: "8LL197pziojWHtS3zeyJonrh1swKvMZpumfesVmDgUcZ"
                  assetsA:
                    - mint: "Asset1MintAddress"
                      isCompressed: true
                    - mint: "Asset2MintAddress"
                      isCompressed: true
                    - mint: "Asset3MintAddress"
                      isCompressed: true
                  assetsB:
                    - mint: "Asset4MintAddress"
                      isCompressed: false
                  solAmountA: "0"
                  solAmountB: "500000000"
      responses:
        '201':
          description: Offer created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      offer:
                        type: object
                        properties:
                          id:
                            type: string
                            example: "1"
                          status:
                            type: string
                            example: "PENDING"
                          makerWallet:
                            type: string
                          takerWallet:
                            type: string
                            nullable: true
                          offeredAssets:
                            type: array
                            items:
                              $ref: '#/components/schemas/Asset'
                          requestedAssets:
                            type: array
                            items:
                              $ref: '#/components/schemas/Asset'
                          offeredSol:
                            type: string
                            description: In lamports
                          requestedSol:
                            type: string
                            description: In lamports
                          createdAt:
                            type: string
                            format: date-time
                      offerType:
                        type: string
                        enum: [ATOMIC, CNFT_BID, BULK_TWO_PHASE]
                        description: |
                          The detected offer type:
                          - `ATOMIC`: Standard swap (NFT↔NFT, NFT↔SOL, cNFT↔SOL)
                          - `CNFT_BID`: cNFT bid with SOL escrow
                          - `BULK_TWO_PHASE`: Bundle swap using lock/settle pattern
                        example: "ATOMIC"
                      transaction:
                        type: object
                        properties:
                          nonceAccount:
                            type: string
                            description: Durable nonce account address
                          message:
                            type: string
                            example: "Transaction will be built when offer is accepted"
                  timestamp:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          description: Validation error - asset ownership or on-chain validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/institution/settings:
    get:
      tags:
        - Institution Settings
      summary: Get client settings
      description: Retrieve the authenticated client's settings including wallets and configuration.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Settings retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/InstitutionSettings'
                  timestamp:
                    type: string
                    format: date-time
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Institution Settings
      summary: Update client settings
      description: Update general settings for the authenticated client (e.g., company name, webhook URL).
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstitutionSettingsUpdate'
            example:
              companyName: "Acme Corporation Pty Ltd"
              webhookUrl: "https://acmecorp.com/webhooks/escrow"
      responses:
        '200':
          description: Settings updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/InstitutionSettings'
                  timestamp:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/institution/clients:
    get:
      tags:
        - Institution Clients
      summary: List clients
      description: List institution clients (paginated). Returns public profile fields only.
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
            maximum: 100
          description: Items per page
        - in: query
          name: includeArchived
          schema:
            type: boolean
            default: false
          description: Include archived clients
      responses:
        '200':
          description: Client list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/InstitutionPublicProfile'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      totalPages:
                        type: integer
                  timestamp:
                    type: string
                    format: date-time
        '401':
          description: Not authenticated
...
components:
...
  schemas:
...
    OfferStatus:
      type: string
      enum:
        - PENDING_INIT
        - PENDING_DEPOSITS
        - SOL_LOCKED
        - NFT_A_LOCKED
        - NFT_B_LOCKED
        - BOTH_LOCKED
        - ALL_LOCKED
        - SETTLED
        - EXPIRED
        - CANCELLED
        - EXPIRED
      description: |
        Agreement status:
        - PENDING_INIT: Agreement is being initialized on-chain
        - PENDING_DEPOSITS: Waiting for required deposits
        - SOL_LOCKED: SOL has been deposited (for NFT<>SOL or NFT+SOL<>NFT swaps)
        - NFT_A_LOCKED: Seller's NFT has been deposited
        - NFT_B_LOCKED: Buyer's NFT has been deposited (for NFT<>NFT swaps)
        - BOTH_LOCKED: Two required assets are locked
        - ALL_LOCKED: All required assets are locked, ready for settlement
        - SETTLED: Agreement has been successfully settled
        - EXPIRED: Agreement has expired without settlement
        - CANCELLED: Agreement has been cancelled
...
  responses:
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for admin authentication
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from /api/v1/institution/auth/login
    SettlementAuthority:
      type: apiKey
      in: header
      name: X-Settlement-Authority-Key
      description: Settlement authority API key for release operations